develop upload file function

+++ video/src/main/java/com/chengxuxiaoba/video/model/Response/VO/UploadFileResponseVo.java
@@ -4,7 +4,7 @@ import java.math.BigInteger;
 
 public class UploadFileResponseVo {
     private String name;
-    private BigInteger size;
+    private Long size;
     private String purpose;
     private String path;
 
@@ -16,11 +16,11 @@ public class UploadFileResponseVo {
         this.name = name;
     }
 
-    public BigInteger getSize() {
+    public Long getSize() {
         return size;
     }
 
-    public void setSize(BigInteger size) {
+    public void setSize(Long size) {
         this.size = size;
     }
